Changed minimax to nagamax variant
The Negamax variant of MinMax is a simplification of the implementation leveraging the fact that the score of a position from your opponent’s point of view is the opposite of the score of the same position from your point of view. So we no longer need to keep track of maximizing player.